+++ test/Test_Example_FG_Range.cpp
@@ -41,7 +41,7 @@ TEST(Example, FG_Range)
   ceres::Solver::Options SolverOptions;
   SolverOptions.trust_region_strategy_type = ceres::TrustRegionStrategyType::DOGLEG;
   SolverOptions.dogleg_type = ceres::DoglegType::SUBSPACE_DOGLEG;
-  SolverOptions.num_threads = std::thread::hardware_concurrency();
+  SolverOptions.num_threads = static_cast<int>(std::thread::hardware_concurrency());
   SolverOptions.minimizer_progress_to_stdout = true;
 
   libRSF::FactorGraph SimpleGraph;
@@ -93,7 +93,7 @@ TEST(Example, FG_Range)
   Expected.addElement(libRSF::Data("point2 1.0 1.0 0.0 0.0 0.0 0.0 0.0"));
   Expected.addElement(libRSF::Data("point2 2.0 1.0 1.0 0.0 0.0 0.0 0.0"));
 
-  /** calculate maximum componentwise absolute difference between solution and expected (mean and covariance) - point2*/
+  /** calculate maximum component-wise absolute difference between solution and expected (mean and covariance) - point2*/
   const double maxAbsErrorMean = libRSF::MaxAbsError(libRSF::DataType::Point2,
                                                      Expected,
                                                      POSITION_STATE,
